moved algorithms to _compare.py
